Move get bound variable list to NodeManager, minor cleanup.
